pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "hindsight2020"
version = "0.1.0"
description = "Hindsight 20/20 — bundles the TradingAgents LangGraph engine (tradingagents + cli packages)."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain-core>=0.3.81",
    "langchain>=0.3.17,<0.4",
    "langfuse>=3.0.0",
    "backtrader>=1.9.78.123",
    "langchain-anthropic>=0.3.15",
    "langchain-experimental>=0.3.4",
    "langchain-google-genai>=2.1.5",
    "langchain-openai>=0.3.23",
    "langgraph>=0.4.8",
    "pandas>=2.3.0",
    "pydantic>=2,<3",
    "parsel>=1.10.0",
    "pytz>=2025.2",
    "questionary>=2.1.0",
    "rank-bm25>=0.2.2",
    "redis>=6.2.0",
    "requests>=2.32.4",
    "rich>=14.0.0",
    "typer>=0.21.0",
    "setuptools>=80.9.0",
    "stockstats>=0.6.5",
    "tqdm>=4.67.1",
    "typing-extensions>=4.14.0",
    "kiteconnect>=5.0.1",
    "yfinance>=0.2.63",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "sse-starlette>=2.1.0",
    "python-dotenv>=1.0.0",
    "supabase>=2.0.0",
]

[project.scripts]
tradingagents = "cli.main:app"

[tool.setuptools.packages.find]
include = ["tradingagents*", "cli*"]
